Enforce the retention policy for AspirEDU backups.
This should look through our past backups and remove any that are not
on a day we'd like to retain.

This adds two commands. One for actually removing backups and one for
creating a fake backup s3 environment to test the command to remove
backups.

This ignores flake8's line length detector. Black handles most of it,
but the long strings aren't auto-shortened. Since black isn't handling
it automatically, let's avoid having to manually adjust it.