Selectively copy backup information from CrunchyBridge.
This revises our approach to only copy today's backup rather than
all ten backups from crunchy bridge to our S3 bucket. It also will
only copy the info file from the archive directory rather than the
entire directory. This is because CB should be using the --copy-archive
option with our backups resulting in the WAL being inserted in the
actual backup directory.

This adds a --dry-run option which will print out summary information
and the commands that will be run.

This now requires the cluster to be specified when running since that
was the default anyway.