Optimize the migration script.
Counting every digit between 000000010001409F000000D4 and 00000001000140A000000007
takes a very long time. The folders only care about the first 16 digits. The rest
of the files use 00-FF in the last two digits.

This also confirms that the backup folder is valid BEFORE trying to fetch
all the relevant folders.